resources/execute/requirements.txt
requests
pymupdf
tqdm
tenacity
numpy
ollama
xinference-client
deepl
openai
azure-ai-translation-text==1.0.1
gradio
huggingface_hub
onnx
onnxruntime
doclayout-yolo==0.0.3
torch==2.5.1
torchvision
opencv-python-headless
tencentcloud-sdk-python
pdfminer.six==20240706
gradio_pdf==0.0.21
pikepdf
peewee==3.17.8
argostranslate
fontTools
python-dotenv
modelscope
flask==2.2.3
flask-cors==3.0.10
gevent==22.10.2
werkzeug==2.2.2